// my-rent-service/src/pages/offer-page/offer-page.tsx
import { JSX,useState } from "react";
// import NearPlaces from "../../components/near-places/near-places";
import {Logo} from "../../components/logo/logo";
import { FullOffer } from "../../types/offer";
import {  useParams } from "react-router-dom";
import ErrorPage from "../error-page/error-page";
import CommentSubmissionForm from "../../components/comment-submission-form/comment-submission-form";
import {ReviewsList} from "../../components/reviews-list/reviews-list";
import { Review } from "../../types/review.ts";
import Map from "../../components/map/map.tsx";
import { CITY } from "../../mocks/city.ts"; 
import { POINTS } from "../../mocks/points.ts";
import MapList from "../../components/mapList/mapList.tsx";
import { Points } from "../../types/map.ts";
import { CitiesCardList } from "../../components/cities-card-list/cities-card-list.tsx";
import { OffersList } from "../../types/offer";

type OfferProps = {
  offers : FullOffer[];
  reviewsList : Review[];
  offersList: OffersList[]
};


function OfferPage({offers,reviewsList,offersList}  : OfferProps): JSX.Element {
    const [selectedPoint, setSelectedPoint] = useState<Points | null>(null);
  
    const handleListItemHover = (listItemName:string) => {
      const currentPoint = POINTS.find((point) =>
        point.title === listItemName,
      );
      setSelectedPoint(currentPoint || null);
    };
  
  
  const params = useParams();
  const offer = offers.find((item) => item.id === params.id)
  if (!offer){
    return <ErrorPage/>
  }
    return(
    <div className="page">
        <header className="header">
          <div className="container">
            <div className="header__wrapper">
              <div className="header__left">
              <Logo/>
              </div>
              <nav className="header__nav">
                <ul className="header__nav-list">
                  <li className="header__nav-item user">
                    <a className="header__nav-link header__nav-link--profile" href="#">
                      <div className="header__avatar-wrapper user__avatar-wrapper">
                      </div>
                      <span className="header__user-name user__name">Myemail@gmail.com</span>
                      <span className="header__favorite-count">3</span>
                    </a>
                  </li>
                  <li className="header__nav-item">
                    <a className="header__nav-link" href="#">
                      <span className="header__signout">Sign out</span>
                    </a>
                  </li>
                </ul>
              </nav>
            </div>
          </div>
        </header>
  
        <main className="page__main page__main--offer">
          <section className="offer">
            <div className="offer__gallery-container container">
              <div className="offer__gallery">
                {offer.images.map((item) =>(
                  <div key={item} className="offer__image-wrapper">
                      <img className="offer__image" src={item} alt="Photo studio"/>
                  </div>
                ))}
              </div>
            </div>
            <div className="offer__container container">
              <div className="offer__wrapper">
                {offer.isPremium ? (
                  <div className="offer__mark">
                  <span>Premium</span>
                </div>) : null
                }
              
                <div className="offer__name-wrapper">
                  <h1 className="offer__name">
                    Beautiful &amp; {offer.title}
                  </h1>
                  <button className="offer__bookmark-button button" type="button">
                    <svg className="offer__bookmark-icon" width="31" height="33">
                      <use href="#icon-bookmark"></use>
                    </svg>
                    <span className="visually-hidden">To bookmarks</span>
                  </button>
                </div>
                <div className="offer__rating rating">
                  <div className="offer__stars rating__stars">
                  <span style={{width:`${(offer.rating / 5) * 100}%` }}></span>
                  <span className="visually-hidden">Rating</span>
                  </div>
                  <span className="offer__rating-value rating__value">{offer.rating}</span>
                </div>
                <ul className="offer__features">
                  <li className="offer__feature offer__feature--entire">
                    {offer.type}
                  </li>
                  <li className="offer__feature offer__feature--bedrooms">
                    {offer.bedrooms} Bedrooms
                  </li>
                  <li className="offer__feature offer__feature--adults">
                    Max {offer.maxAdults} adults
                  </li>
                </ul>
                <div className="offer__price">
                  <b className="offer__price-value">&euro;{offer.price}</b>
                  <span className="offer__price-text">&nbsp;night</span>
                </div>
                <div className="offer__inside">
                  <h2 className="offer__inside-title">What&apos;s inside</h2>
                  <ul className="offer__inside-list">
                    {offer.goods.map((item) =>(
                      <li key={item} className="offer__inside-item">
                        {item}
                      </li>
                    ))}
                  </ul>
                </div>
                <div className="offer__host">
                  <h2 className="offer__host-title">Meet the host</h2>
                  <div className="offer__host-user user">
                    <div className="offer__avatar-wrapper offer__avatar-wrapper--pro user__avatar-wrapper">
                      <img className="offer__avatar user__avatar" src={offer.host.avatarUrl} width="74" height="74" alt="Host avatar"/>
                    </div>
                    <span className="offer__user-name">
                      {offer.host.name}
                    </span>
                    {offer.host.isPro ? (
                    <span className="offer__user-status">
                      Pro
                    </span>) : null
                    } 
                  </div>
                  <div className="offer__description">
                    <p className="offer__text">
                      {offer.description}
                    </p>
                    <p className="offer__text">
                      An independent House, strategically located between Rembrand Square and National Opera, but where the bustle of the city comes to rest in this alley flowery and colorful.
                    </p>
                  </div>
                </div>
                <section className="offer__reviews reviews">
                  <h2 className="reviews__title">Reviews &middot; <span className="reviews__amount">1</span></h2>
                    <ReviewsList reviewsList={reviewsList}/>
                  <CommentSubmissionForm/>

                </section>
              </div>
            </div>
            <section className="offer__map">
            <h1>Парки города {CITY.title}:</h1>
                <MapList points={POINTS} onListItemHover={handleListItemHover}/>
                <Map city={CITY}
                points={POINTS}
                  selectedPoint={selectedPoint}
                  />
            </section>
          </section>
          <div className="container">
            <section className="near-places places">
              <h2 className="near-places__title">Other places in the neighbourhood</h2>
              <div className="near-places__list places__list">
              <CitiesCardList offersList={ offersList }/>
              </div>
            </section>
          </div>
        </main>
      </div>
    );
}


export default OfferPage;